move bool, date and dateRangeFacet into filters
