+++ test/example-types/filters/dateRangeFacet.js
@@ -1,7 +1,7 @@
 let _ = require('lodash/fp')
 let { expect } = require('chai')
-let { getDateIfValid, rollingRangeToDates } = require('../../src/utils/dateUtil')
-let dateRangeFacet = require('../../src/example-types/dateRangeFacet')
+let { getDateIfValid, rollingRangeToDates } = require('../../../src/utils/dateUtil')
+let dateRangeFacet = require('../../../src/example-types/filters/dateRangeFacet')
 
 let commonFilterParts = {
   type: 'dateRangeFacet',
